+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Bugfix in `getBody` method returning `null`.
+- Bugfix in `addRequest` method not JSON-encoding data.
 
 ## [3.0.0] - 2025.01.15